requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
motor==3.6.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
pymongo==4.9.2
pytest==8.4.2
python-dotenv==1.0.1
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
stripe>=11.0.0,<15.0.0
typing_extensions==4.15.0
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
slowapi==0.1.9
pytest-asyncio==0.24.0


